+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Group changes to describe their impact on the project, as follows:
 | Known Issues   | Issues whose fix has not been tested and cannot be included in this release |
 
 ## [2.5.1]
+### [Fixed]
+- **Logger:** Some useful error logs were not printed during the startup phase.
+
 ### [Removed]
 - **Proxy:**
   - **Global:**